requirements.txt
Django==5.0.1
python-dotenv==1.0.0
groq>=0.9.0
openai>=1.0.0
huggingface_hub>=0.20.0
requests>=2.31.0,<3.0.0
pdfplumber>=0.10.3
python-docx>=0.8.11
SQLAlchemy>=2.0.30
pyodbc>=5.1.0
mssql-django>=1.4
spacy>=3.7.0,<3.8.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
# Note: If you encounter pydantic compatibility issues with Python 3.12,
# you may need to install: pip install "pydantic<2.0" or upgrade spacy

